update(ProxyRefreshParser): converts DFC-v1 non-LDP lists to use LDP:Container
DFC v1 does not require data-servers to send us their data in containers, which causes problems with RDFLib's graph parser and DjangoLDP's serializer. Our workaround is to transform the lists of string IDs we receive into proper containers. This is achieved with implicit knowledge of the DFC fields, and a recursive search of the data received.